Add files via upload
These are the newer data files for the updated version of the smallmouth bass manuscript. Changes include reformatting content for updated analyses, removal of two fish that were not treated consistently from all analyses, and other minor updates noted in the "Notes" column of each sheet where applicable and/or in the actual manuscript for this project. Older data files (and raw data) are still available upon request!